polish MQProtos.h
